requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
ecdsa==0.19.0
idna==3.7
pyasn1==0.6.0
python-dotenv==1.0.1
python-jose==3.3.0
requests==2.32.2
rsa==4.9
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
